+++ m7/src/Utils/shared_memory.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <Arduino.h>
-#include "CalibrationData.h"
+#include "Utils/CalibrationData.h"
 
 #define CHAR_BUFFER_SIZE 256
 #define FLOAT_BUFFER_SIZE 256
@@ -20,7 +20,7 @@ struct FloatCircularBuffer {
 };
 
 struct VoltageCircularBuffer {
-  float buffer[VOLTAGE_BUFFER_SIZE];
+  double buffer[VOLTAGE_BUFFER_SIZE];
   volatile uint32_t read_index;
   volatile uint32_t write_index;
 };
@@ -63,7 +63,7 @@ bool m4HasCharMessage();
 
 bool m4SendFloat(const float* data, size_t length);
 
-bool m4SendVoltage(const float* data, size_t length);
+bool m4SendVoltage(const double* data, size_t length);
 
 bool m7SendChar(const char* data, size_t length);
 bool m7ReceiveChar(char* data, size_t& length);
@@ -72,5 +72,5 @@ bool m7HasCharMessage();
 bool m7ReceiveFloat(float* data, size_t& length);
 bool m7HasFloatMessage();
 
-bool m7ReceiveVoltage(float* data, size_t& length);
+bool m7ReceiveVoltage(double* data, size_t& length);
 bool m7HasVoltageMessage();
